Update dependencies and Dockerfile versions
This commit updates the base images in the Dockerfile to use golang:1.25-alpine3.23 and alpine:3.23. Additionally, it upgrades several dependencies in go.mod and go.sum, including:
- github.com/bavix/features to v1.0.4
- github.com/go-playground/validator/v10 to v10.28.0
- github.com/goccy/go-yaml to v1.19.0
- google.golang.org/grpc to v1.77.0
- google.golang.org/protobuf to v1.36.10

Furthermore, it updates the golangci-lint version in the Makefile and various GitHub Actions workflows to ensure compatibility with the latest features and improvements.